Convert Lockin data to array
For some reason, the lockin data (magdata) was being passed as a tuple,
and causing the program to crash.  Now we convert it to a numpy array
before using it.  I also changed the defaults to reflect the number we
are currently using, I was tired of changing it all the time while
debugging!