Automatically searches for resonance
Changed the DPPH measurement to search for resonances automatically;
this prevents the user from forgetting to press the “find resonance”
button; if the range is wrong, then the result will look terrible and
it will force the user to enter a new range and search for the
resonance again.  Also, I cleaned up the status messages of dpph_utils
a bit.